updated entrypoint, readme, terraform and circleci

+++ scripts/terraform/terraform_destroy.sh
@@ -39,8 +39,14 @@ then
     echo "Terraform directory not set. Using default directory."
     /root/.local/bin/aws sts get-caller-identity
     cd "$TERRAFORM_ROOT"
-    /usr/local/bin/terraform init -input=false
-    /usr/local/bin/terraform destroy -auto-approve
+    if [ "${TERRAFORM_DESTROY_ALTERNATE_COMMAND}" == "true" ]; then
+        TERRAFORM_COMMAND=$(shyaml get-value terraform_options.terraform_apply.command < bitops.config.yaml || true)
+        exec "${TERRAFORM_COMMAND}"
+    else
+        /usr/local/bin/terraform init -input=false
+        /usr/local/bin/terraform plan
+        /usr/local/bin/terraform destroy -auto-approve
+    fi
 fi  
 
 printf "${SUCCESS} Successfully destroyed Terraform deployment..."
